requirements.txt
# NAVO Requirements - Simplified Teams Integration

# Core web framework
fastapi==0.115.13
uvicorn==0.32.1

# Microsoft Bot Framework for Teams
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0

# HTTP client for API calls
aiohttp==3.11.11

# OpenAI for Enterprise GPT
openai==1.58.1

# Environment and configuration
python-dotenv==1.0.1

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0

